corerequirements.txt
pillow
camelot-py
boto3>=1.12.0,<1.13.0
constants
botocore
wagtail
python3_ghostscript
PyPDF4
drf-yasg
django-rest-swagger
social-auth-app-django==4.0.0
opencv-python
xmltodict
python-social-auth
unipath
django-ckeditor
emails
django-social-share